requirements.txt
dspy-ai>=2.6.24
openai>=1.82.0
numpy>=2.2.0
scikit-learn>=1.6.0
pandas>=2.2.0
chromadb>=1.0.12
sentence-transformers>=3.0.0
python-dotenv>=1.1.0
tqdm>=4.67.0
litellm>=1.71.0
optuna>=4.3.0
ujson>=5.10.0
tenacity>=9.1.0
pyyaml>=6.0.2
torch>=2.6.0
transformers>=4.51.0
tiktoken>=0.9.0